Windows port changes. Windows has overloaded sqrt so casting to double removes the ambiguity.  The MS compiler also needed more explicit class and namespace declararations in a few cases.  The winmath.h file was included in a few places to define some missing functions from math.h

+++ src-std-mesh/Sources/SundancePartitionedRectangleMesher.cpp
@@ -31,7 +31,7 @@ PartitionedRectangleMesher::PartitionedRectangleMesher(const ParameterList& para
 
 void PartitionedRectangleMesher::balanceXY(int n, int* npx, int* npy)
 {
-  int m = (int) floor(sqrt(n));
+  int m = (int) floor(sqrt((double)n));
   for (int i=m; i>=1; i--)
   {
     if (n % i == 0)
